feedback-service/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
psycopg2-binary
python-dotenv
kafka-python
requests